Update and rename myBB_GetTypeDB.rb to mybb_get_type_db.rb
